Remove some styles occurrences
