+++ dockerize/docker/REQUIREMENTS.txt
@@ -46,7 +46,6 @@ drf-yasg~=1.21
 django-rest-multiple-models==2.1.3
 
 django-preferences==1.0.0
-PyWavefront==1.3.3
 geoip2==4.5.0
 django-matomo==0.1.6
 uwsgi~=2.0
