+++ qgis-app/templates/base.html
@@ -1,5 +1,5 @@
 {% load i18n simplemenu_tags static %}
-{% load resources_custom_tags %}
+{% load plugin_utils %}
 {% load matomo_tags %}
 <!DOCTYPE html>
 <html xmlns="http://www.w3.org/1999/xhtml">
